Refactor cli module to smaller files
The file `cli.py` had 1518 LOC effectively making it the largest file in
the project. This PR splits cli module into submodules commands, parsers
and move utils inside the cli.

Also, the main module had supressed warnings as:
- C901 : Function is too complex
- PLR0912 : too many branches
- PLR0915 : too many statements
