+++ src/datachain/cli/commands/__init__.py
@@ -1,0 +1,29 @@
+from .datasets import (
+    dataset_stats,
+    datasets,
+    edit_dataset,
+    list_datasets_local,
+    rm_dataset,
+)
+from .du import du
+from .index import index
+from .ls import ls
+from .misc import clear_cache, completion, garbage_collect
+from .query import query
+from .show import show
+
+__all__ = [
+    "clear_cache",
+    "completion",
+    "dataset_stats",
+    "datasets",
+    "du",
+    "edit_dataset",
+    "garbage_collect",
+    "index",
+    "list_datasets_local",
+    "ls",
+    "query",
+    "rm_dataset",
+    "show",
+]
